rag for interview/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-faiss>=0.0.1
openai>=1.5.0
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
tiktoken>=0.5.1
ollama>=0.1.5
langsmith>=0.0.62